refactor: refine UserProfile app files structure
