+++ ui/packages/tidb-dashboard-lib/src/apps/UserProfile/components/index.ts
@@ -1,0 +1,4 @@
+export * from './Form.SSO'
+export * from './Form.Session'
+export * from './Form.PrometheusAddr'
+export * from './Form.Version'
